split contentconverter
